Fix incorrect of TempDir instance in a context manager.
In the previous implementation of TempDir, entering a second
context manager will create a second temporary directory. I believe
this does not meet most user's expectations, which is that they wish
to change the CWD to the newly created temporary directory.

```python
    >>> from path import TempDir
    >>> with TempDir() as td:  # Create the temporary directory.
    ...     with td:  # Enter the temporary directory... right?
    ...               # Wrong! It created another temporary directory!
    ...
```

This commit changes the return value of the `__enter__` dunder method
from `self` to `self._next_class(self)`, which has the effect of
making the above code example work as expected.