Copy annotations from __init__.pyi to __init__.py
__init__.pyi has been removed.

Annotations were copied from the stuf file as-is, without changing
either the annotations or anything about the source code (with
exceptions mentioned below). In some cases, the annotations are not
correct (or better ones could be made) or mypy will now flag code inside
the annotated methods - these will be corrected in subsequent commits.
This commit is intended to serve as a reference point for translation
from stub annotations to in-source annotations.

The exceptions to as-is copy are the following:
- When the annotation for an input parameter was "Path | str" it has
  been replaced with "str" since "Path" is a subclass of "str"
- When the return type was "Path" it has been replaced with "Self"
  (except in the cases where it truely should be "Path" like in
  "TempDir")
- In cases where the stub contained explicit parameters and the source
  had "*args, **kwargs", these have been replaced with explicit
  parameters to match the stub. In several of these cases, this required
  changing the source code to call the forwarded function with the
  explict parameters.