Replace Number with float or int
The Number ABC is more abstract than needed when defining allowed input
types - float captures both float and int.

Further, when returning explictly an int, the return type should be
annotated as such instead of as Number.

This also eliminates a mypy error about multiplying a Number with a
float.