Ensure shutil-derived functions are properly typed
When assigning a shutil function directly into the Path namespace, the
typing system seems to have trouble figuring out that the first argument
is the implict self, and as a result the type checking fails for these
functions.

To solve, thin passthrough functions have been added so that the
annotations are correct for the Path context, but the docstrings have
been copied from shutil as-is.